refactor(broker): extract SessionsStorage seam from SessionsBroker
Phase 4 (#2978): sessions on Postgres. SessionsBroker becomes a thin
wrapper delegating to an injected SessionsStorage, mirroring how
EventBroker/MemoryBroker delegate to their Stream<T> implementation
instead of owning storage logic directly. The current mutate-in-place
logic moves unchanged into InMemorySessionsStorage; the pure per-query
merge rules (resolveQueryPhase, updateExistingQuery, buildQueryEntry)
move into a shared session-aggregate module so a future Postgres
backend can reuse them verbatim instead of duplicating the merge
semantics.

Every method on the seam is now async, since a Postgres-backed
implementation will need to await a locked read-modify-write. Route
handlers, the events/messages fan-out, and graceful shutdown are
updated to await accordingly, and now thread the triggering event's
or message's sequence number through - unused by the in-memory
backend, but a Postgres backend needs it as an idempotency watermark.

No behavior change: the in-memory backend is the only implementation
so far, and its logic is a straight lift.
